Combine Error types
