--- src/types/ErrorMsg.ts
@@ -1,6 +1,0 @@
-export enum ErrorMsg {
-  AddError = 'Unable to add a todo',
-  UpdateError = 'Unable to update a todo',
-  DeleteError = 'Unable to delete a todo',
-  NoError = '',
-}
