--- src/types/SetError.ts
@@ -1,3 +1,0 @@
-import { ErrorMsg } from './ErrorMsg';
-
-export type SetError = (err?: boolean, msg?: ErrorMsg) => void;
